Feature ETP-3981: F13 — remove dead businessPartner dir, move shared chart to contacts
