+++ docs/generated-custom-windows/contacts.md
@@ -125,7 +125,7 @@ The Contacts window should let users maintain a shared business-partner master r
 - `tools/app-shell/src/components/contract-ui/PartnerAddressPicker.jsx` and `tools/app-shell/src/components/contract-ui/__tests__/PartnerAddressPicker.test.js` confirm that the same contacts location modal now supports inline "+ Add address" creation for partner-address selectors outside the Contacts window.
 - `tools/app-shell/src/components/contract-ui/CreateContactModal.jsx` provides partial supporting evidence for person/company create payload semantics in the shared quick-create flow, but no contacts-window-specific automated test was found for the main detail-route save behavior.
 - `tools/app-shell/src/menu.json` and `tools/app-shell/src/windows/registry.js` confirm menu visibility and route-to-loader registration for `/contacts`.
-- `tools/app-shell/src/windows/custom/contacts/BusinessPartnerSidebar.jsx` confirms the contacts-specific sidebar implementation: KPI cards with color-coded revenue/expenses, a localized trend chart built on the shared `BPChartSVGContent` exported from `businessPartner/BusinessPartnerSidebar`, and an expandable dialog with a 3M/6M period toggle.
+- `tools/app-shell/src/windows/custom/contacts/BusinessPartnerSidebar.jsx` confirms the contacts-specific sidebar implementation: KPI cards with color-coded revenue/expenses, a localized trend chart built on the local `BPChartSVGContent` (in `contacts/BPChartSVGContent.jsx`), and an expandable dialog with a 3M/6M period toggle.
 - `tools/app-shell/src/components/contract-ui/__tests__/DetailView.autoSaveOnBlur.test.js` is a static regression guard (ETP-3660) that reads `DetailView.jsx` as a string and asserts: `autoSaveOnBlur` prop has a `false` default; `handleFieldBlur` exists alongside `hook.editing`, `hook.selected`, and `hook.handleSave` usage; `onFieldBlur={autoSaveOnBlur ? handleFieldBlur : undefined}` appears on both Form instances (principal and collapsed sections).
 - `modules/com.etendoerp.go/src-test/src/com/etendoerp/go/schemaforge/ContactNameSyncHandlerTest.java` contains 14 JUnit 5 unit tests (>80 % coverage) for `ContactNameSyncHandler`: happy paths (first-name-only change, last-name-only, both changed, trimming, null partial names), isPerson guards (false, null), no-changes guard, blank-name guard, isValidEvent guards (TriggerHandler disabled, wrong entity), and getObservedEntities caching. Uses `MockedStatic<ModelProvider>` and `MockedStatic<TriggerHandler>`; resets the static `entities` field via reflection in `@BeforeEach`.
 - `tools/app-shell/src/lib/apiError.js` is the shared utility for parsing non-ok fetch Responses into human-readable error messages. It supports NEO Headless format (`{error: {message}}`), Etendo JsonDataService format (`{response: {error: {message|string}}}`), top-level `{message}`, and falls back to `Error ${status}` for non-JSON bodies. Used by both `ContactsTable.jsx` (single-row delete) and `contacts/index.jsx` (bulk delete).
